tools/packagebuilder/requirements.txt
asyncio-nats-client==0.10.0
asyncio-nats-streaming==0.4.0